#87 Add interfaces commons
