+++ interfaces/telegram/__init__.py
@@ -1,10 +1,1 @@
-bot_instance = None
 
-
-def __init__(bot):
-    global bot_instance
-    bot_instance = bot
-
-
-def get_bot():
-    return bot_instance
